requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
fastapi==0.68.1
h11==0.9.0
httptools==0.1.2
idna==3.2
kafka-python==2.0.2
pydantic==1.8.2
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.11.3
uvloop==0.16.0
vk-api==11.9.4
websockets==8.1
